Upgraded to Bullet 2.71 Minor changes to Falling Cubes demo: use 6 static planes, and clamp delta time. Use btDbvtBroadphase by default
git-svn-id: http://oolongengine.googlecode.com/svn@77 7fa7efda-f44a-0410-a86a-c1fddb4bcab7

+++ trunk/Oolong Engine2/Bullet/src/LinearMath/btStackAlloc.h
@@ -29,7 +29,7 @@ struct btBlock
 	unsigned char*		address;
 };
 
-///StackAlloc provides some fast stack-based memory allocator (LIFO last-in first-out)
+///The StackAlloc class provides some fast stack-based memory allocator (LIFO last-in first-out)
 class btStackAlloc
 {
 public:
@@ -61,7 +61,7 @@ public:
 
 	int	getAvailableMemory() const
 	{
-		return totalsize - usedsize;
+		return static_cast<int>(totalsize - usedsize);
 	}
 
 	unsigned char*			allocate(unsigned int size)
